Module1 complete project as per following requirements:
Add code to the file so that your applet does the following (see the hints that follow for help if you need it):

Display a second map next to the default one. This map should be zoomed into your hometown by default -- see hints below for how to find your hometown latitude and longitude. Use a different Map Provider for each of the two maps. If you are working offline, you should not change the map provider, but you should use the “dark” tiles which are in the file “blankLight-1-3.mbtiles.” For some Map Provider choices for working online, see the hints below.
Change the background of the display to be a different color. BONUS: can you include a visual effect in the background (shading / gradient)?