As part of module4 assignment added code as per following requirements:
1) Implement the isLand(Feature earthquake) method in EarthquakeCityMap. This method should return true if
the location of the input earthquake is on land. It should also set the "country" property on the LandMarker
to the country where the earthquake occurred. Otherwise, the location is in the ocean and the method should return false.

2) Implement the printQuakes() method in EarthquakeCityMap. This method should use System.out.println() to list
each country for which there was 1 or more earthquakes and the number of earthquakes detected in that country.
Then it should print out the number of quakes that were detected in the ocean. Note that this method is not trivial.
You will have to calculate the number of earthquakes per country from the information you have available. As an aside:
If you are running the applet with a large earthquake file/feed (e.g. 1.0+ Past week or 30 days), you might find that
printQuakes takes a long time to run.

3) Complete the method definition for the draw() method in the CityMarker class. We suggest you use triangles to
represent cities, but of course the exact shape and color is up to you. Then add the marker to your key (which
you will need to modify as you change your earthquake markers). Your map should now look something like the image below.

4)  Complete the method definitions for the drawEarthquake() method in LandQuakeMarker and OceanQuakeMarker so that
the size of the marker varies by its magnitude (larger magnitude=larger marker). In addition, make the LandQuakeMarkers
one shape (we suggest circles) and the OceanQuakeMarkers a different shape (we suggest squares). Be sure to center
them at the x, y position passed in to drawEarthquake. Also, they will appear white for now. That’s OK. Do not set their
color in drawEarthquake. That will be done in the next step. Then add your markers to the key. Your map should now
look something like the image below. Make sure there is a visual difference between the size of the markers for
different magnitudes.

5) Complete the method definition for colorDetermine() method in EarthquakeMarker and notice where it and the methods
from the derived classes are called in the EarthquakeMarker’s draw method.

6)  Modify the draw method so that it draws an X over all earthquake markers whose earthquakes have occurred in the last day.