Module 5 project as per following requirements:
1)  Implement the selectMarkerIfHover helper method in EarthquakeCityMap. This method is called in
mouseMoved() method in EarthquakeCityMap, and mouseMoved() is called by the event handler when the
user moves the mouse. selectMarkerIfHover should set the instance variable selected for the first
Marker it finds that mouseX and mouseY is inside of.

2)  Implement the showTitle(PGraphics pg, float x, float y) in both EarthquakeMarker and CityMarker
so that it displays the relevant information about the city or earthquake. For earthquake markers
display the title of earthquake. For a city, display its name, country, and population.

3) Implement the mouseClicked() method in EarthquakeCityMap. This method tests whether the lastClick
variable refers to a null object or not; if lastClick is not null, this click “de-selects” whichever
marker was clicked on last and so all markers should be displayed.