Modul6 project requirements completed:
1) Implement the selectMarkerIfHover helper method in EarthquakeCityMap. This method is called in mouseMoved() method in EarthquakeCityMap, and mouseMoved() is called by the event handler when the user moves the mouse. selectMarkerIfHover should set the instance variable selected for the first Marker it finds that mouseX and mouseY is inside of.

2)  Implement the showTitle(PGraphics pg, float x, float y) in both EarthquakeMarker and CityMarker so that it displays the relevant information about the city or earthquake. For earthquake markers display the title of earthquake. For a city, display its name, country, and population.

3) Implement the mouseClicked() method in EarthquakeCityMap. This method tests whether the lastClick variable refers to a null object or not; if lastClick is not null, this click “de-selects” whichever marker was clicked on last and so all markers should be displayed. Otherwise, the method must determine which marker is being selected (if any) and hide / display other markers so that:

When an earthquake’s marker is selected, all cities within the threat circle of this earthquake are displayed on the map and all other cities and earthquakes are hidden. You are given an implementation of the threatCircle() method in the EarthquakeMarker class.

When a city’s marker is selected, all earthquakes which contain that city in their threat circle are displayed on the map and all other cities and earthquakes are hidden. You are given an implementation of the threatCircle() method in the EarthquakeMarker class.