Assignment: How Easy to Read is Your Writing?
1) getNumSentences()
This method counts and returns the number of sentences in the document. A sentence is any sequence of characters
ending in one or more end of line punctuation marks, or the last sequence of characters in a document, even if
they don’t end with an end of line punctuation mark. An end of line punctuation mark includes a period (.), an
exclamation point (!) and a question mark (?).

2) getNumWords()

This method counts and returns the number of words in the document (the stored text). A word is any contiguous
sequence of one or more alphabetical characters. We will completely ignore numbers when we count words, and you
can assume you will not have any strings that combine numbers and letters

3) getNumSyllables()

This method counts and returns the total number of syllables in the document (the stored text). To count the number
of syllables in a word, you should use the following rule:

Each contiguous sequence of one or more vowels is a syllable, with the following exception: a lone "e" at the end of
a word is not considered a syllable unless the word has no other syllables. You should consider y a vowel.

Under these rules the words “the”, “fly”, “yes”, “cave” and “double” all have 1 syllable, but "segue" has two syllables.
Notice that this isn’t exactly correct (“double” actually has 2 syllables), but it’s close enough for our purposes.
Here are some more examples with the number of syllables your method should return to help you: "contiguous" (3 syllables),
"sleepy" (2 syllables), "obvious" (2 syllables), "toga" (2 syllables). Notice that our rules get a lot wrong, especially
when you have more than 2 vowels in a row, but these are the rules we will test you against.