Assignment: Spelling Suggestions
1) public void insertions(String s, List<String> currentList, boolean wordsOnly)

This method behaves just like substitute above, except it tries to insert any letter in between
existing letters (or to the start and end) to create new Strings (which it adds to currentList).

2) public void deletions(String s, List<String> currentList, boolean wordsOnly)

This method behaves just like substitute above, except it tries to remove any existing letter
to create new Strings (which it adds to currentList).

3) public List<String> suggestions(String word, int numSuggestions)

This method allows us to implement the SpellingSuggest Interface and is very similar to the Breadth
First Search you performed with AutoComplete in Module 4. In this method, we'll be looking at
providing a number of spelling suggestions for a misspelled word.

4) To avoid huge search introduce THRESHOLD or number after finding those many words stop search