Assignment: Spell Checking and Autocomplete
1. Implement the DictionaryLL class which implements the Dictionary interface using a LinkedList data structure.
For this assignment, we are ignoring the case of words. Your dictionary should store all words as lower case words,
i.e. they should be converted to lower case before you put them in. Similarly, your dictionary should convert a word
to lower case before it checks to see whether it is in the Dictionary. You will have the option to enable case sensitivity
(e.g. so "Christine" is considered a word, but "christine" or "CHristine" is flagged as misspelled) in the optional
extension to this project, described below.

Test your class by running the JUnit test suite we provide. Note: there is very little to do to implement this class so
if you think it's too easy, you're probably doing it right.

2. Implement the DictionaryBST class which implements the Dictionary interface using a TreeSet (balanced Binary Search Tree)
data structure. For this assignment, we are ignoring the case of words. Your dictionary should store all words as lower case
words, i.e. they should be converted to lower case before you put them in. Similarly, your dictionary should convert a word
to lower case before it checks to see whether it is in the Dictionary. You will have the option to enable case
sensitivity (e.g. so "Christine" is considered a word, but "christine" or "CHristine" is flagged as misspelled) in the
optional extension to this project, described below.

3. Implement the missing methods in AutoCompleteDictionaryTrie.java to use a trie data structure to store the words in a
dictionary. Each method is described in more detail below.

As in part 1, your implementation should ignore case. That is, you should convert each word to lowercase before inserting
it into the trie. When you look for a word in the trie, you should look for the all lowercase version of the word. The
methods Character.toLowerCase and/or String.toLowerCase will come in handy here.

Your implementation should make use of the TrieNode class we have provided for you. You should not need to modify this
class at all, but if you want to you may add methods to this class, but you should not remove or modify any of these methods.

public boolean addWord(String word)

This method adds a word by creating and linking the necessary trie nodes into the trie, as described outlined in the
videos for this week. It should appropriately use existing nodes in the trie, only creating new nodes when necessary.
E.g. If the word "no" is already in the trie, then adding the word "now" would add only one additional node (for the 'w').

public int size()

Returns the number of valid words stored in the dictionary. Note this is NOT necessarily the same as the number of
TrieNodes in the trie.

public boolean isWord(String s)

Return whether or not the given string is in the trie, using the algorithm described in the videos for this week.

public List<String> predictCompletions(String prefix, int numCompletions)

Return a list containing the numCompletions shortest legal completions of the prefix string. All legal completions must
be valid words in the dictionary. If the prefix itself is a valid word, it should be included in the list of returned
words. The list of completions must contain all of the shortest completions, but when there are ties, it may break them
in any order. For example, if there the prefix string is "ste" and only the words "step", "stem", "stew", "steer" and
"steep" are in the dictionary, when the user asks for 4 completions, the list must include "step", "stem" and "stew",
but may include either the word "steer" or "steep".