Assignment: Implement and Test a Linked List
1.
A)
Author the following methods in MyLinkedList.java:

public MyLinkedList()

This constructor should create the Linked List and setup any instance variables as needed.

public E get(int index)

This method returns the node in the list corresponding to the index. For example, if the list has the elements 6 and 9
(in that order) and you call "get" with index 0, you should get back 6. If the method is called with an invalid index
(say -1 or 2 in the example above), you will thrown an IndexOutOfBoundsException.

public boolean add( E element )

This method adds an element to the end of the list. Be sure to think about what happens when you add an element to an
empty list as well as a list with already existing elements. Drawing a picture of the list will help!

B)
Author the "testAddEnd" and "testAddAtIndex" methods in MyLinkedListTester

The tests for "testGet" may not properly stress the add method. To be sure your add method works properly, you should
add tests to these two methods.

2. Implement and write tests for the remaining methods in MyLinkedList.

public int size()

This method returns the size of the list. Be careful to update your size instance variable when adding and removing elements.
To test this method, use the method in MyLinkedListTester called testSize

public void add(int index, E element)

This method adds an element at the specified index. If elements exist at that index, you will move elements at that index (and
beyond) up, effectively inserting this element at that location in the list. Although drawings are helpful for implementing any
method, you will benefit heavily from drawing out what should happen in this method before trying to code it. You will want to
throw an IndexOutOfBoundsException if the index provided is not reasonable for inserting an element.
Optional: After authoring this version of the add method, you could consider removing redundant code by having the add method you
originally wrote just call this method. To test this method, use the method in MyLinkedListTester called testAddAtIndex.

public E remove(int index)

This method removes an element from that location. Similar to the "add" method, we strongly encourage you to draw this out before
trying to code it. Like the "get" method, you will want to thrown an IndexOutOfBoundsException if the index provided is out of
bounds. To test this method, use the method in MyLinkedListTester called testRemove.

public E set(int index, E element)

This method allows you to change elements in place by specifying a valid index of an existing element and a new value to change
that element to. You should return the old value held at that index. Like the "get" method, you will want to thrown an IndexOutOfBoundsException
if the index provided is out of bounds. In addition, if someone tries to set a node with a null element, you should throw a NullPointerException.
To test this method, use the method in MyLinkedListTester called testSet.