Assignment: Implementing Graphs
Part 1: Implement the degreeSequence method

The degree sequence of a graph is a sorted (organized in numerical order from largest to smallest, possibly with
repetitions) list of the degrees of the vertices in the graph.

The following two methods are already provided in the starter code:

getNeighbors(int v) an abstract method in Graph.java that is implemented in both derived classes. When given the
input vertex index v, this method returns a list containing the integer indices of vertices appearing as the end
point of edges starting at v. Note: if there are multiple edges between v and one of its neighbors, then that neighbor
appears multiple times in this list.

getInNeighbors(int v) similar to above except returning list of indices of vertices which are the start points of
edges that end at v.

Part 2: Implement the getDistance2 method in both GraphAdjList and GraphAdjMatrix implementations

In this step of the assignment, you'll implement a method which takes the name of a node in the graph and returns the
list of nodes that can be reached in two hops. Just like the getNeightbors( ) and getInNeighbors( ) methods, the
implementation of the abstract method must go in the classes GraphAdjList and GraphAdjMatrix.

public List<Integer> getDistance2(int v)

To facilitate auto-grading, you need to read the method descriptions carefully to ensure they have the correct behavior.
Also, make sure you are consistent with our test cases/examples in the main method.

In particular, if there are multiple two-hop paths to get to a vertex, then that vertex should be listed multiple times.