Assignment: Project: Class Design and Graph Search
1)
Complete the implementation of MapGraph

Your goal in this assignment is to complete the implementation of the MapGraph class, implementing
all of the methods described below. (Some of the methods in MapGraph will be implemented next week).

In completing this implementation, we expect you do to any or all of the following:

Add additional classes to the roadgraph package
Add additional methods (usually private, but perhaps public as well) to the MapGraph class
Add member variables to the MapGraph class

2)
The required methods are:
public MapGraph() Creates a new empty MapGraph object.
public boolean addVertex(GeographicPoint location) Adds a vertex to the graph at the location specified
by the GeographicPoint. Each vertex in the graph represents an intersection at a unique geographic point.
Returns true if the vertex was successfully added, and false if it was not (because the vertex representing
location was already in the graph) or if the parameter object is null.

public void addEdge(GeographicPoint end1, GeographicPoint end2, String roadName, String roadType, double length)
Adds a directed edge between two GeographicPoints (lat, lon pairs) that are already in the graph. The edges
represent streets segments, and hence they connect intersections. The roadName is the name of the road
(e.g. "Main street") while the roadType is the kind of road (e.g. "residential"). The length is the length
of this road segment, in km. This method should throw an IllegalArgumentException if either of the two points are
not in the graph already, if any of the arguments is null, or if length is less than 0.

public int getNumVertices() returns the number of vertices in the graph
public Set<GeographicPoint> getVertices() returns a set of the vertices in the graph in terms of their
GeographicPoint (i.e. the locations of all of the intersections). Hint: A HashSet is a good choice for your set.

public int getNumEdges() returns the number of edges in the graph
public List<GeographicPoint> bfs(GeographicPoint start, GeographicPoint goal,Consumer<GeographicPoint> nodeSearched)
Performs bfs starting at start until it reaches goal and returns a list of geographic points along the shortest
(unweighted) path from start to goal. This list should include both the start and the goal points and should be
ordered from start to goal.

public List<GeographicPoint> bfs(GeographicPoint start, GeographicPoint goal) Creates a dummy Consumer object
and calls your bfs method. You do NOT need to modify this method. Useful for testing!

MapGraph also defines the following methods which you should NOT implement this week.
public List<GeographicPoint> dijkstra(GeographicPoint start, GeographicPoint goal,Consumer<GeographicPoint> nodeSearched)
Performs dijkstra's algorithm to search the graph starting at the start until it reaches the goal and returns a list of
geographic points along the shortest (unweighted) path from start to goal. You do NOT need to implement this method this week.

public List<GeographicPoint> aStarSearch(GeographicPoint start, GeographicPoint goal, Consumer<GeographicPoint> nodeSearched)
Performs the A-Star algorithm to search the graph starting at the start until it reaches the goal and returns a list of
geographic points along the shortest (unweighted) path from start to goal. You do NOT need to implement this method this week.